enhancement: 외부 API 호출 모듈 분리 (news, suggestion, trend)
