+++ app/request_external_api/headers.py
@@ -1,9 +1,5 @@
 import os
 
-NAVER_API_HEADERS = {
-    'X-Naver-Client-Id': os.getenv('NAVER_API_CLIENT_ID'),
-    'X-Naver-Client-Secret': os.getenv('NAVER_API_CLIENT_SECRET'),
-}
 
 CLOVA_API_HEADERS = {
     "X-NCP-APIGW-API-KEY-ID": os.getenv('CLOVA_API_CLIENT_ID'),
