--- app/request_headers/__init__.py
@@ -1,3 +1,0 @@
-from app.request_headers.headers import CLOVA_API_HEADERS
-from app.request_headers.headers import CLOVA_SUMMARY_API_ENDPOINT
-from app.request_headers.headers import NAVER_API_HEADERS
