+++ main.py
@@ -1,6 +1,5 @@
 import json
 import os
-import re
 from datetime import datetime, timedelta
 
 import pandas as pd
@@ -13,10 +12,10 @@ from starlette.templating import Jinja2Templates
 
 from app.util.completion_executor import CompletionExecutor
 from app.util.news_comments_crawler import NewsCommentsCrawler
-from app.request_headers.headers import CLOVA_API_HEADERS
-from app.request_headers.headers import CLOVA_SUMMARY_API_ENDPOINT
-from app.request_headers.headers import NAVER_API_HEADERS
 from app.util.string_util import StringUtil
+from app.request_external_api.request_news import RequestNews
+from app.request_external_api.request_suggestions import RequestSuggestions
+from app.request_external_api.request_trend import RequestTrend
 load_dotenv()
 
 app = FastAPI()
@@ -24,50 +23,16 @@ app.mount("/static", StaticFiles(directory="static"), name="static")
 templates = Jinja2Templates(directory="templates")
 
 
-def get_naver_news(query, display, start=1, sort='date'):
-    naver_news_api_url = f'https://openapi.naver.com/v1/search/news.json?query={query}&display={display}&start={start}&sort={sort}'
-    return requests.get(naver_news_api_url, headers=NAVER_API_HEADERS)
 
-
-def get_clova_summary_result(content, tone=0, summary_count=1):
-    # tone - 0: 원문 어투, 1: 해요체 2: 정중체, 3: 명사형 종결체
-    data = {
-        "document": {
-            "content": content
-        },
-        "option": {
-            "language": "ko",
-            "model": "general",
-            "tone": tone,
-            "summaryCount": summary_count
-        }
-    }
-
-    try:
-        response = requests.post(CLOVA_SUMMARY_API_ENDPOINT, headers=CLOVA_API_HEADERS, data=json.dumps(data))
-        response.raise_for_status()
-        return response.json()['summary']
-    except requests.exceptions.HTTPError as http_err:
-        print(f"HTTP error occurred: {http_err}")  # HTTP 에러 출력
-        print(f"Response content: {response.content.decode()}")  # 응답 본문 출력
-        return None
-    except Exception as err:
-        print(f"Other error occurred: {err}")  # 기타 에러 출력
-
-
-def get_suggestions(query):
-    suggestions_api_url = f"http://suggestqueries.google.com/complete/search?client=firefox&ds=yt&q={query}&hl=ko"
-    suggestions_response = requests.get(suggestions_api_url)
-    return suggestions_response.json()[1]
 
 
 def get_today_issue_summary(q: str):
     news_title = []
 
-    suggestions = get_suggestions(q)
+    suggestions = RequestSuggestions.get_suggestions(q)
 
     for suggestion in suggestions[:1]:
-        naver_news_response = get_naver_news(suggestion, display=10, sort='sim')
+        naver_news_response = RequestNews.get_naver_news(suggestion, display=10, sort='sim')
         if naver_news_response.status_code != 200:
             raise HTTPException(status_code=500, detail="NEWS API 호출 오류")
 
@@ -104,42 +69,15 @@ def get_today_issue_summary(q: str):
     }
     return completion_executor.execute(request_data).split('\n')
 
-
-def get_trend_searh_data(start_date: str, end_date: str, time_unit: str, keyword_groups: list):
-    # 구간단위 - date, week, month
-    # group data => groupName, keywords:list
-    # 날짜 형식 yyyy-mm-dd
-    naver_trend_search_api_endpoint = "https://openapi.naver.com/v1/datalab/search"
-
-    request_body = {
-        "startDate": start_date,
-        "endDate": end_date,
-        "timeUnit": time_unit,
-        "keywordGroups": keyword_groups
-    }
-
-    try:
-        response = requests.post(naver_trend_search_api_endpoint, headers=NAVER_API_HEADERS,
-                                 data=json.dumps(request_body))
-        response.raise_for_status()
-        return response.json()['results']
-    except requests.exceptions.HTTPError as http_err:
-        print(f"HTTP error occurred: {http_err}")  # HTTP 에러 출력
-        print(f"Response content: {response.content.decode()}")  # 응답 본문 출력
-        return None
-    except Exception as err:
-        print(f"Other error occurred: {err}")  # 기타 에러 출력
-
-
 def get_trend_variation(q: str):
     today = datetime.today()
     two_months_ago = (today - timedelta(days=60)).replace(day=1)
     two_months_ago = two_months_ago.strftime('%Y-%m-01')
     today = today.strftime('%Y-%m-%d')
 
-    keyword_groups = [{'groupName': q, 'keywords': [suggestion for suggestion in get_suggestions(q)]}]
+    keyword_groups = [{'groupName': q, 'keywords': [suggestion for suggestion in RequestSuggestions.get_suggestions(q)]}]
 
-    trend_search_data = get_trend_searh_data(two_months_ago, today, 'date', keyword_groups)[0]['data']
+    trend_search_data = RequestTrend.get_naver_trend_search_data(two_months_ago, today, 'date', keyword_groups)[0]['data']
     daily_variation = (trend_search_data[-1]['ratio'] / trend_search_data[-2]['ratio'] * 100) - 100
     two_weeks_ago_ratio = sum([entry['ratio'] for entry in trend_search_data[-14:-7]])
     one_weeks_ago_ratio = sum([entry['ratio'] for entry in trend_search_data[-7:]])
@@ -177,9 +115,9 @@ def get_suggestion_trend(q: str):
     today = datetime.today()
     one_week_ago = (today - timedelta(days=7)).strftime('%Y-%m-%d')
     today = today.strftime('%Y-%m-%d')
-    suggestions = get_suggestions(q)
+    suggestions = RequestSuggestions.get_suggestions(q)
     keyword_groups = [{'groupName': suggestion, 'keywords': [suggestion]} for suggestion in suggestions[1:6]]
-    return get_trend_searh_data(one_week_ago, today, 'date', keyword_groups)
+    return RequestTrend.get_naver_trend_search_data(one_week_ago, today, 'date', keyword_groups)
 
 
 def get_suggestion_trend_score(trend_data: list):
@@ -219,36 +157,13 @@ def get_suggestion_entire_data(q: str):
     return suggestion_entire_data
 
 
-def convert_news_url_to_comment_url(news_url):
-    # 뉴스 URL의 정규식 패턴
-    pattern = r"(https://n\.news\.naver\.com/mnews/article/)(\d+/\d+)(\?.*)"
-    replacement = r"\1comment/\2\3"
-
-    # 정규식을 사용하여 댓글 URL로 변환
-    comment_url = re.sub(pattern, replacement, news_url)
-
-    return comment_url
-
-
-def extract_first_two_parentheses_content(text):
-    # 정규 표현식으로 첫 번째와 두 번째 괄호 안의 내용 추출
-    pattern = r'\(([^)]+)\)'
-    matches = re.findall(pattern, text)
-
-    # 첫 번째와 두 번째 괄호 안의 내용만 반환
-    if len(matches) >= 2:
-        return matches[:2]
-    else:
-        return matches
-
-
 def get_comment_sentiment_data(q: str):
     comment_sentiment_data = []
 
-    suggestions = get_suggestions(q)
+    suggestions = RequestSuggestions.get_suggestions(q)
     news_link = []
     for suggestion in suggestions[:1]:
-        naver_news_response = get_naver_news(suggestion, display=10, sort='sim')
+        naver_news_response = RequestNews.get_naver_news(suggestion, display=10, sort='sim')
         if naver_news_response.status_code != 200:
             raise HTTPException(status_code=500, detail="NEWS API 호출 오류")
 
@@ -261,7 +176,7 @@ def get_comment_sentiment_data(q: str):
     news_comments_crawler = NewsCommentsCrawler()
 
     for link in news_link:
-        tmp = news_comments_crawler.parse(convert_news_url_to_comment_url(link))
+        tmp = news_comments_crawler.parse(StringUtil.convert_news_url_to_comment_url(link))
         if tmp:
             raw_data.append(tmp)
 
@@ -301,7 +216,7 @@ def get_timeline(q: str):
     all_articles = []
 
     for i in range(1, 2):
-        all_articles += get_naver_news(q, 100, i, 'sim').json()['items']
+        all_articles += RequestNews.get_naver_news(q, 100, i, 'sim').json()['items']
     for article in all_articles:
         title = article['title']
         link = article['link']
